backend/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
numpy==2.3.3
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
SQLAlchemy==2.0.43
threadpoolctl==3.6.0
twilio
google-generativeai==0.3.2
python-dotenv==1.0.0
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
